requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-dotenv==1.0.0
pywhatkit==5.4
eventlet==0.33.3
gevent==23.9.1
gevent-websocket==0.10.1
gunicorn==21.2.0
whitenoise==6.6.0
